+++ crates/api/src/widget_base/mod.rs
@@ -32,11 +32,41 @@ pub fn toggle_flag(flag: &str, widget: &mut WidgetContainer) {
 
     if let Some(selector) = widget.try_get_mut::<Selector>("selector") {
         if value {
-            selector.set_state(flag);
+            selector.push_state(flag);
         } else {
-            selector.clear_state();
+            selector.remove_state(flag);
         }
     }
+}
+
+/// Sets the given property flag to  `true` and set the flags name as state of the widgets selector.
+pub fn set_flag(flag: &str, widget: &mut WidgetContainer) {
+    if !widget.has::<bool>(flag) {
+        return;
+    }
+
+    widget.set(flag, true);
+
+    if let Some(selector) = widget.try_get_mut::<Selector>("selector") {
+        selector.push_state(flag);
+    }
+
+    widget.update(false);
+}
+
+/// Sets the given property flag to  `false` and clears the state of the widgets selector.
+pub fn remove_flag(flag: &str, widget: &mut WidgetContainer) {
+    if !widget.has::<bool>(flag) {
+        return;
+    }
+
+    widget.set(flag, false);
+
+    if let Some(selector) = widget.try_get_mut::<Selector>("selector") {
+        selector.remove_state(flag);
+    }
+
+    widget.update(false);
 }
 
 /// Used to define the `parent_type`of a widget.